IOS-2903 fixes: test not building independently and dependently to my changes

+++ SubProjects/MessageModel/MessageModel/MessageModelTests/TestUtils/TestUtil.swift
@@ -153,7 +153,7 @@ extension TestUtil {
         //SMTP
         let createe = CdServer(context: moc)
         createe.serverType = serverType
-        createe.port = Int16(serverPort)
+        createe.port = Int32(serverPort)
         createe.address = serverAddress
         createe.transport = serverTransport
         createe.automaticallyTrusted = automaticallyTrusted
